reorganizeComponent
