+++ src/pages/ExpensesPage.js
@@ -1,5 +1,5 @@
 import React from "react";
-import LogInForm from "../pages/LogInForm";
+import LogInForm from "../components/LogInForm";
 function ExpensesPage() {
   return <LogInForm />;
 }
